crams-apps/crams_storage/requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
merc_common==1.0.0
crams_contact==1.0.0
crams-log==1.0.0
mysqlclient
rest_condition==1.0.3